+++ datapack_development/data/scripts/handlers/itemhandlers/FishShots.java
@@ -69,7 +69,7 @@ public class FishShots implements IItemHandler
 		
 		int FishshotId = item.getItemId();
 		int grade = weaponItem.getCrystalType();
-		int count = item.getCount();
+		long count = item.getCount();
 		
 		if ((grade == L2Item.CRYSTAL_NONE && FishshotId != 6535) || (grade == L2Item.CRYSTAL_D && FishshotId != 6536) || (grade == L2Item.CRYSTAL_C && FishshotId != 6537) || (grade == L2Item.CRYSTAL_B && FishshotId != 6538)
 				|| (grade == L2Item.CRYSTAL_A && FishshotId != 6539) || (grade == L2Item.CRYSTAL_S && FishshotId != 6540) || (grade == L2Item.CRYSTAL_S80 && FishshotId != 6540))
